backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
stripe>=10.0.0
python-multipart>=0.0.9
httpx>=0.27.0

# Dev / Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
